% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importgcam.R
\name{addMIBatchCSV}
\alias{addMIBatchCSV}
\title{Parse the GCAM ModelInterface output (DEPRECATED)}
\usage{
addMIBatchCSV(fn, proj, clobber = FALSE, transformations = NULL,
  saveProj = TRUE, saveProjEach = FALSE)
}
\arguments{
\item{fn}{Name of the file containing the output from the GCAM Model
Interface.}

\item{proj}{Project to add extracted results to.  Can be either a project
data structure or the name of a project data file.  The file will be created
if it doesn't already exist.}

\item{clobber}{If \code{TRUE}, overwrite any existing scenario of the same
name; otherwise, fail if scenario/query already exists in the data set.}

\item{transformations}{Transformation functions to apply to the queries (see
details).}

\item{saveProj}{A flag to save the project to disk after data has been added.
A user may want to avoid it if they are for instance calling this method several
times and would prefer to save at the end.  Users can always save at anytime by
calling \code{saveProject}.}

\item{saveProjEach}{A flag to save the project to disk after each query has
completed. This would be useful if a user suspects a failure in the middle
of running queries and would like to not loose progress made.}
}
\value{
The project dataset with the new scenario added.
}
\description{
Parse the raw output of a GCAM batch query and add them into a project.
}
\details{
This function could be usefult for users who have exsisting CSV batch query
output they would like to import without having to re-run the queries.
}
