% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importmisc.R
\name{addQueryTable}
\alias{addQueryTable}
\title{Import a table of data into a project data set as a new query.}
\usage{
addQueryTable(project, qdata, queryname, clobber = FALSE,
  strict.rundate = FALSE)
}
\arguments{
\item{project}{The project data set to add the query to.  This can be either
a project data structure the name of a project data file.}

\item{qdata}{Query data to add.  This can be either a data frame or the name
of a file containing the data in csv format.}

\item{queryname}{The name to use for this query when it is stored in the
project.}

\item{clobber}{Flag: if \code{TRUE}, then the operation can replace a query
in an existing scenario from the data set.  If \code{FALSE}, then attempting
to replace an existing query will cause the entire operation to fail.}

\item{strict.rundate}{Flag: if \code{TRUE}, then require that the run dates
match in order to add queries to a scenario, and fail the entire operation if
they don't.  If \code{FALSE}, then ignore dates in the new data set.}
}
\description{
This function allows you to add a free-standing table to your project data.
Unlike \code{\link{addScenario}}, it does not attempt to run the GCAM Model
Interface; therefore, it is suitable for cases where the data has come from
some other source, such as the output of a post-run analysis code.
}
\details{
The data to be read can be either a data frame or the name of a file
containing the data.  In the latter case, the data will be read using
\code{read.csv}.  The format of the table should be similar to the format of
tables produced by the Model Interface.  Namely:
\itemize{
  \item There should be a "scenario" column with the scenario name in it.  The
date of the run, if any, will be stripped off.
  \item There should be one or more columns that identify each data point,
such as "region" or "sector".  Column names will be converted to lower case.
  \item There should be a "Units" column that gives the measurement unit for
the data.
  \item The results for model years should appear in the remaining columns,
one for each year.  For historical reasons, most utilities for handling GCAM
data expect these columns to start with 'X', for example: "X2050"; however,
this convention is not enforced here.
}

You can have multiple scenarios in a single table if you want.  In that case
the relevant data will be added to each scenario that appears in the
"scenario" column.  However, because the run date is stripped off, the
scenarios must have distinct scenario names.  Having multiple copies of the
same scenario (i.e., the same scenario name, but different run dates) is
\emph{not} supported, and attempting to load such a table will result in a
single query with duplicated rows.

The scenarios found will be added to the project data set supplied.  This can
be either the data set itself, or the name of a file containing the data
set.  Either way, the modified data set will be written back to the file it
came from.  If the attempt to add the query to \emph{any} scenarios fails,
then the function will throw an error and the project data will be
unmodified.
}

