% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importgcam.R
\docType{data}
\name{DEFAULT.MICLASSPATH}
\alias{DEFAULT.MICLASSPATH}
\title{Default java class path for running the Model Interface}
\format{An object of class \code{character} of length 1.}
\usage{
DEFAULT.MICLASSPATH
}
\description{
This is the class path you get if you pass \code{NULL} to any of the
functions that run (directly or indirectly) the model interface.  The
necessary jar files are bundled with the package, so unless you need a
customized version of the model interface for some reason, this should take
care of all your GCAM query needs.
}
\keyword{datasets}

