% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangledata.R
\name{diffScenarios}
\alias{diffScenarios}
\title{Compute a difference between two scenarios}
\usage{
diffScenarios(data, scenarios = NULL, keycols = "region",
  data2 = NULL, aggfun = sum)
}
\arguments{
\item{data}{The input query data.  The expected format is described below.}

\item{scenarios}{The scenarios to difference from the input data.  Order is
significant; it determines which scenario will be subtracted, and which will
be subtracted from.}

\item{keycols}{Columns by which to aggregate the data.  The default is to
aggregate by region.  If \code{NULL}, then don't do any aggregation.}

\item{data2}{If the two scenarios are in separate data frames, pass the
scenario to be subtracted in this argument.  In this case, the
\code{scenarios} argument will be ignored.}

\item{aggfun}{Function to use in aggregation.  The default is \code{sum}.}
}
\value{
Data frame with the difference between the two scenarios.  The key
columns and year columns will be retained; all other columns (including the
\code{scenario} column will be dropped.
}
\description{
This function takes a single query for two scenarios and computes the
difference between them for all years in the scenario. Entries that exist in
one of the scenarios but not the other will be assigned a NA value in the
output.
}
\details{
The input data should consist of a data frame with a \code{scenario} column,
one or more id columns (\code{region}, \code{sector}, etc.), and a series of
columns containing the data.  We assume that any column with numeric data in
it is a data column.  Any columns other than the key columns and data columns
will be dropped as part of the aggregation.

By default, both scenarios are expected to be contained in a single data
frame, distinguished by the \code{scenario} column.  There can be more than
two scenarios in the data frame; the \code{scenarios} argument will determine
which ones actually get used.  The \code{scenarios} argument also determines
the order of the differencing (i.e., \code{scenario1 - scenario2}
vs. \code{scenario2 - scenario1}.

By default the data is aggregated by region before differencing, but this
behavior can be changed with the \code{keycols} argument.  You can specify
one or more key columns to change the aggregation variable, or you can set
\code{keycols=NULL} to retain all of the input columns and not do any
aggregation.  Aggregation is by summing, by default, but an alternative
aggregation function can be specified in the \code{aggfun} argument.

If you happen to have the data for the two scenarios in separate data frames,
you can pass the second data frame in the \code{data2} argument.  In this
case, the \code{scenarios} argument will be ignored, and the difference will
be computed as \code{data - data2}.
}
