% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageProjectData.R
\name{dropScenarios}
\alias{dropScenarios}
\title{Remove a scenario from a project data set}
\usage{
dropScenarios(proj, scenarios, invert = FALSE,
  writeback = is.character(proj))
}
\arguments{
\item{proj}{Project data or data file name.}

\item{scenarios}{Name(s) of the scenario(s) to drop}

\item{invert}{If \code{TRUE} then delete all scenarios \emph{except} the ones
given in \code{scenarios}.}

\item{writeback}{If \code{TRUE} then write the change into the backing data
file.}
}
\description{
This function removes one or more scenarios from a project data set and
returns the newly modified data.  If this operation is performed on a file,
the file is also modified.  If it is performed on a data set that has already
been loaded, then the change is passed onto the file only if \code{writeback}
is set to TRUE.
}
\details{
Technically is is possible to pass a filename for \code{proj} with
\code{writeback} set to \code{FALSE}.  This usage will cause the data set to
be loaded and filtered without modifying the original, albeit in a somewhat
nonintuitive way.
}

