% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageProjectData.R
\name{getQuery}
\alias{getQuery}
\title{Retrieve a query for one or more scenarios}
\usage{
getQuery(projData, query, scenarios = NULL)
}
\arguments{
\item{projData}{The data set to extract from.}

\item{query}{The name of the query to extract.}

\item{scenarios}{Vector of scenario names.  If NULL, use all scenarios in the
data set.}
}
\description{
Return a data frame with the results for a query for all of the selected
scenarios.
}
