% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageProjectData.R
\name{getRundates}
\alias{getRundates}
\title{Return the run date for one or more scenarios in a data set.}
\usage{
getRundates(projData, scenarios = NULL)
}
\arguments{
\item{projData}{The data set to report on.}

\item{scenarios}{The names of the scenarios}
}
\description{
The run dates are recorded in the tables produced by the GCAM Model Interface
and copied into the project data set on import.  This function retrieves the
run dates for the selected scenarios and returns them as a named vector.
}

