% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageProjectData.R
\name{listQueries}
\alias{listQueries}
\title{List the queries available for a scenario}
\usage{
listQueries(projData, scenarios = NULL, anyscen = TRUE)
}
\arguments{
\item{projData}{The data set to report on.}

\item{scenarios}{The name(s) of the scenario(s) to report on.  If NULL,
report on all of them.}

\item{anyscen}{If \code{TRUE}, then list queries that are in any scenario.
If \code{FALSE}, list queries that are in all scenarios.}
}
\description{
Return the names of the queries available for a scenario in a project data
set.  Unlike \code{\link{listScenarios}}, this function requires the data set
to have been previously loaded, so it cannot take a file name.
}
