% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageProjectData.R
\name{listScenarios}
\alias{listScenarios}
\title{List the scenarios in a project data set}
\usage{
listScenarios(projData)
}
\arguments{
\item{projData}{The data set to report on.}
}
\description{
Return the names of the scenarios available in a project data set.  The input
can be either the name of a file containing the data set, or the data
structure previously loaded from such a file.
}

