% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querymi.R
\name{listScenariosInDB}
\alias{listScenariosInDB}
\alias{listScenariosInDB.localDBConn}
\alias{listScenariosInDB.remoteDBConn}
\title{Lists the Scenarios contained in a GCAM Database}
\usage{
listScenariosInDB(dbConn)

\method{listScenariosInDB}{localDBConn}(dbConn)

\method{listScenariosInDB}{remoteDBConn}(dbConn)
}
\arguments{
\item{dbConn}{The connection to a database which will handle running the query.}
}
\value{
A table with the results.
}
\description{
To run a query users need to supply a database connection, the query to run,
the scenarios to query, and regions to query.
}
\section{Methods (by class)}{
\itemize{
\item \code{localDBConn}: Run a query on a local GCAM database

\item \code{remoteDBConn}: Run query specialization for remote databases
}}

