% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querymi.R
\name{localDBConn}
\alias{localDBConn}
\title{Create a connection to a local database that can be use to run queries on.}
\usage{
localDBConn(dbPath, dbFile, miclasspath = NULL)
}
\arguments{
\item{dbPath}{The path in which the BaseX DBs are located.}

\item{dbFile}{GCAM database to extract scenario from.}

\item{miclasspath}{Java class path for the GCAM Model Interface.}
}
\value{
A connection to a local BaseX databasse which can be used to run
queries.
}
\description{
In order to establish a so called connection, really we track the location
files as well) which is the classpath TODO: link to documentation and syntax
In addition we require the path in which the databases reside and (TODO:
optionally?) the name of the database to query.
}
