% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageProjectData.R
\name{mergeProjects}
\alias{mergeProjects}
\title{Merge a list projects into a single project}
\usage{
mergeProjects(prjname, prjlist, clobber = FALSE, saveProj = TRUE)
}
\arguments{
\item{prjname}{The name of a project data file for the merged project.}

\item{prjlist}{A list of projects that need to be merged together.  Note
each project will be run through \code{\link{loadProject}} in case it has
not yet been loaded.}

\item{clobber}{If \code{TRUE}, overwrite any existing scenario of the same
name; otherwise, fail if scenario/query already exists in the data set.}

\item{saveProj}{A flag to save the project to disk after data has been added.
A user may want to avoid it if they are for instance calling this method several
times and would prefer to save at the end.  Users can always save at anytime by
calling \code{saveProject}.}
}
\value{
The project dataset with the projects merged.
}
\description{
Users can use this function to collapse multiple project into a single
project.  This could be useful for instance to import someone else's
data into your project.  A user must explicitly specify the new name
the new project will recieve.  In addition they can control what
happens with scenario/query collisions with the \code{clobber} param.
}
