% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querymi.R
\name{miquery_post}
\alias{miquery_post}
\title{Post-process raw query results}
\usage{
miquery_post(results, query, scenarios, regions, warn.empty)
}
\arguments{
\item{results}{Table returned by the ModelInterface query.}

\item{query}{The original query string.}

\item{scenarios}{The scenarios requested in the query.}

\item{regions}{The regions requested in the query.}

\item{warn.empty}{Flag: issue warning if the results table is empty.}
}
\description{
The results returned by the ModelInterface query are not aggregated (if, for
instance, the query is by region).  This function performs that aggregation,
if applicable.
}
\keyword{internal}
