% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importgcam.R
\name{parse_batch_query}
\alias{parse_batch_query}
\title{Parse a Model Interface batch query file so each aQuery can get sent to be
run individually.}
\usage{
parse_batch_query(fn)
}
\arguments{
\item{fn}{Name of the batch query file to parse}
}
\value{
A list of queries by name of a list $regions and $query which has
the list of regions to query and the query text.
}
\description{
Parse a Model Interface batch query file so each aQuery can get sent to be
run individually.
}
\keyword{internal}
