% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importgcam.R
\name{parse_mi_output}
\alias{parse_mi_output}
\title{Parse the GCAM ModelInterface output (DEPRECATED)}
\usage{
parse_mi_output(fn)
}
\arguments{
\item{fn}{Name of the file containing the output from the GCAM Model
Interface.}
}
\description{
Parse the raw output of a GCAM batch query into a set of tables.
Normally you will pass the structure returned by this function to
\code{process_batch_q} extract (and optionally do some filtering
and processing on) the table you are looking for.
}
\details{
This function is obsolete and is included only for backward compatibility
with old code that uses it.  Instead, use \code{\link{addScenario}} to
process GCAM data into a native structure.
}

