% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importgcam.R, R/manageProjectData.R
\name{project.valid}
\alias{project.valid}
\title{Check whether a project data structure is valid.}
\usage{
project.valid(prjdata)

project.valid(prjdata)
}
\arguments{
\item{prjdata}{The data structure to test.}

\item{prjdata}{The data structure to test.}
}
\description{
This function will check the 7 conditions described in
\code{\link{loadProject}}.  It will return 0 if they pass, or the number of
the first condition to fail otherwise.

This function will check the 7 conditions described in
\code{\link{loadProject}}.  It will return 0 if they pass, or the number of
the first condition to fail otherwise.
}

