% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importgcam.R
\docType{data}
\name{PROTOTYPE.MIBATCH}
\alias{PROTOTYPE.MIBATCH}
\title{Prototype Model Interface batch file}
\format{An object of class \code{character} of length 14.}
\usage{
PROTOTYPE.MIBATCH
}
\description{
This is a prototype for the batch file that drives the Model Interface.  It
has a bunch of tags like [SCENARIO], [DBFILE], etc. for
\code{\link{runModelInterface}} to fill in with the particulars of the
queries we are trying to run.
}
\keyword{datasets}

