% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamgmt.R
\name{query.valid}
\alias{query.valid}
\title{Check whether a query table in a scenario component structure is valid.}
\usage{
query.valid(querytable)
}
\arguments{
\item{querytable}{The query table to test.}
}
\description{
This function will check conditions 5-7 for a single query table in a
scenario data structure.  It will return the number of the first check to
fail, or 0 if all pass.
}

