% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querymi.R
\name{remoteDBConn}
\alias{remoteDBConn}
\title{Create a connection to a remote database that can be use to run queries on.}
\usage{
remoteDBConn(
  dbFile,
  username,
  password,
  address = "localhost",
  port = 8984,
  validatedb = TRUE
)
}
\arguments{
\item{dbFile}{GCAM database to extract scenario from.}

\item{username}{A username configured with READ access on the remote BaseX
database server.}

\item{password}{The password for the said username. WARNING: currently just
stored and sent as plain text, does BaseX even support https?}

\item{address}{The server address such as IP or domain name address.  Default
is "localhost"}

\item{port}{The server port.  The default is 8984, the same as the default
used by BaseX.}

\item{validatedb}{If \code{TRUE}, check that a simple db query works on the
connection; otherwise, don't run the check.}
}
\value{
A connection to a remote BaseX databasse which can be used to run
queries.
}
\description{
This connection will attempt to connect to a remote BaseX server via a
webserver using the BaseX REST API.  In order to connect a user must supply
the server address, port, username/password as configured in the BaseX
server which as read access.
In addition we require TODO: optionally?) the name of the database to query.
}
