% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangledata.R
\name{rsub}
\alias{rsub}
\title{Perform a difference reduction}
\usage{
rsub(...)
}
\arguments{
\item{...}{The values to be reduced}
}
\value{
The difference reduction of the values, with special cases handled as
described in details.
}
\description{
This is the subtraction equivalent of sum(...).  Specifically, rsub(x1, x2,
..., xn) is equivalent to x1 - sum(x2, ... xn).  However, we make one
exception:  if only a single argument is passed, we return NA.  This is so
that the result makes sense in \code{\link{diffScenarios}}.
}
\keyword{internal}
