% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importgcam.R
\name{run_batch_queries}
\alias{run_batch_queries}
\title{Run a parsed list of batch queries on the given database connection and
for the given scenario.}
\usage{
run_batch_queries(conn, scenarios, query_list)
}
\arguments{
\item{conn}{A database connection.}

\item{scenarios}{A list of scenario to query.}

\item{query_list}{A list of queries to run as generated from \code{\link{parse_batch_query}}.}
}
\value{
A list of data tables representing the query results.
}
\description{
Run a parsed list of batch queries on the given database connection and
for the given scenario.
}
\keyword{internal}
