% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importgcam.R
\name{runModelInterface}
\alias{runModelInterface}
\title{Run the GCAM Model Interface with results placed into a temporary file.}
\usage{
runModelInterface(dbFile, scenario = NULL, queryFile = NULL,
  miclasspath = NULL, migabble = NULL)
}
\arguments{
\item{dbFile}{The GCAM output database to run queries on.}

\item{scenario}{The name of the scenario to query.  If \code{NULL}, then take
the last one in the database.}

\item{queryFile}{Name of the file containing the queries.  If \code{NULL},
then use the built-in default.}

\item{miclasspath}{Classpath for the GCAM model interface.  If \code{NULL},
then use the built-in default.}

\item{migabble}{Control what happens to Model Interface console output.}
}
\description{
This function will run the Model Interface on the selected GCAM output
database.  The results will be placed in a temporary file, the name of which
will be returned from the function.
}

