% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querymi.R
\name{runQuery.localDBConn}
\alias{runQuery.localDBConn}
\title{Run query specialization for local databases}
\usage{
\method{runQuery}{localDBConn}(dbConn, query, scenarios, regions)
}
\description{
Run query specialization for local databases
}
