% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querymi.R
\name{runQuery.remoteDBConn}
\alias{runQuery.remoteDBConn}
\title{Run query specialization for remote databases}
\usage{
\method{runQuery}{remoteDBConn}(dbConn, query, scenarios, regions)
}
\description{
Run query specialization for remote databases
}
