% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importgcam.R
\name{SAMPLE.GCAMDB}
\alias{SAMPLE.GCAMDB}
\title{Sample GCAM database file.}
\usage{
SAMPLE.GCAMDB()
}
\description{
This file can be used for testing the mechanics of running GCAM queries using
the functions in this package.  A lot of the data normally included in the
database has been stripped out to control the total size, so a lot of queries
won't work properly.  The \code{\link{SAMPLE.QUERIES}} should all work.
}
