% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importgcam.R
\docType{data}
\name{SAMPLE.QUERIES}
\alias{SAMPLE.QUERIES}
\title{Sample GCAM query file}
\format{An object of class \code{character} of length 1.}
\usage{
SAMPLE.QUERIES
}
\description{
These queries should work with the sample data in
\code{\link{SAMPLE.GCAMDB}}.
}
\keyword{datasets}

