% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageProjectData.R
\name{scenario.valid}
\alias{scenario.valid}
\title{Check whether a scenario component in a project data structure is valid.}
\usage{
scenario.valid(scendata)
}
\arguments{
\item{scendata}{The data structure to test}
}
\description{
This function will check conditions 3-7 for a single scenario component in a
project data structure.  It will return the number of the first check to
fail, or 0 if all pass.
}

