% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamgmt.R
\name{sep.date}
\alias{sep.date}
\title{Separate the scenario and date in GCAM scenario tags.}
\usage{
sep.date(scenstr)
}
\arguments{
\item{scenstr}{The scenario string.  This may be a vector, such as the
'scenario' column in a table of GCAM results}
}
\value{
\code{list(scenario=Scenario_strings, date=Dates)}
}
\description{
GCAM scenarios are named according to a user-supplied name and the date the
scenario was run.  The resulting format is
"Scenario_Name,date=YYYY-DD?-MM?THH:MM:SS-hh:mm".  ('X?' represents an X that might
or might not be present.  The user string and date are separated by exactly
one space.)  This function, therefore, splits the name string from the date.
The scenario name is returned as is, and the date is converted to a date
object.  The results are returned as a list with elements 'scenario' and
'date'.
}
\details{
This function assumes that it can split the
}

