% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importgcam.R
\name{stdcase}
\alias{stdcase}
\title{Standardize the case of a table's columns.}
\usage{
stdcase(tbl)
}
\arguments{
\item{tbl}{The table to standardize}
}
\description{
GCAM is a little inconsistent about the case of the column names in its
tables.  Usually it's all lower case, but occasionally we get a 'Region'
instead of 'region'.  To keep analysis functions from having to guess at the
case they need to be looking for we standardize to all lower case, except for
the XYYYY year columns, and 'Units', which existing tools all expect to be in
title case.
}
\keyword{internal}
