% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamgmt.R
\name{table.scen.trim}
\alias{table.scen.trim}
\title{Trim the 'date=' from the scenario column in a table}
\usage{
table.scen.trim(tbl)
}
\description{
Return a version of a GCAM results table in which the scenario name contains
just the name of the scenario, without the date information that is typically
packed into that column.
}

