% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageProjectData.R
\name{listQueries}
\alias{listQueries}
\title{List the queries available for a scenario}
\usage{
listQueries(projData, scenario)
}
\arguments{
\item{projData}{The data set to report on.}

\item{scenario}{The name of the scenario to report on.}
}
\description{
Return the names of the queries available for a scenario in a project data
set.  Unlike \code{\link{listScenarios}}, this function requires the data set
to have been previously loaded, so it cannot take a file name.
}

