% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageProjectData.R
\name{loadProject}
\alias{loadProject}
\title{Load a project data file.}
\usage{
loadProject(proj)
}
\arguments{
\item{proj}{Project to add extracted results to.  Can be either a project
data structure or the name of a project data file.  The file will be created
if it doesn't already exist.}
}
\value{
The project dataset which may or may not have acutally had to be loaded.
}
\description{
This function will load a project data file and return the project data
structure.  Project files can be created with \code{\link{addScenario}}.
}
\details{
Project data is stored as an R 'save' file with a single variable called
'prjdata'.  If the data file doesn't have the expected variable in it, the
function will throw an error.  It also does some minimal checks to ensure
that the data is valid, but there is a limit to what it can do.
Specifically, it checks to see that the structure loaded is:
    1. a list,
    2. with at least one element, and
    3.   all of those elements are lists,
    4.   with at least one element, and
    5.     all of those elements are data frames,
    6.     with a scenario column, and
    7.     at least one other column.
If the validation fails, the error message will indicate which of these steps
it failed on.

Despite these checks, it is possible to construct a data set that passes and
yet still contains bad data.  When in doubt load the file directly and check
to see that it contains the data you expect it to.
}
