% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querymi.R
\name{localDBConn}
\alias{localDBConn}
\title{Create a connection that can be used to run queries on a local GCAM database}
\usage{
localDBConn(dbPath, dbFile, miclasspath = NULL, migabble = TRUE,
  validatedb = TRUE, maxMemory = "4g")
}
\arguments{
\item{dbPath}{The path in which the BaseX DBs are located.}

\item{dbFile}{GCAM database to extract scenario from.}

\item{miclasspath}{Java class path for the GCAM Model Interface.}

\item{migabble}{If \code{TRUE}, discard model interface console output.  If
\code{FALSE}, display console output.}

\item{validatedb}{If \code{TRUE}, check that a simple db query works on the
connection; otherwise, don't run the check.}

\item{maxMemory}{Sets the maximum memory for Java which will be used to run
the queries.  The default value is \code{"4g"}.  Users may need to reduce this
value if they are using a 32-bit Java or increase it if they suspect they are
running out of memory (by using \code{migabble} to check the log).  Note the
numeric value can be suffixed with "g" for Gigabyte or "m" for Megabyte.}
}
\value{
A connection to a local BaseX databasse which can be used to run
queries.
}
\description{
Given a directory in which a GCAM database is located and the name of the
database, return a connection that can be used to run queries on the
database.
}
\details{
By default, the a version of the GCAM ModelInterface and BaseX libraries
supplied with the package will be used to run the query.  You can replace
these by specifying a Java classpath minimally including the replacement
ModelInterface.jar and BaseX.jar files.
}
