% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querymi.R
\name{localDBConn}
\alias{localDBConn}
\title{Create a connection to a local database that can be use to run queries on.}
\usage{
localDBConn(dbPath, dbFile, miclasspath = NULL, migabble = NULL)
}
\arguments{
\item{dbPath}{The path in which the BaseX DBs are located.}

\item{dbFile}{GCAM database to extract scenario from.}

\item{miclasspath}{Java class path for the GCAM Model Interface.}

\item{migabble}{Control what happens to the model interface console output.
Default is to discard.}
}
\value{
A connection to a local BaseX databasse which can be used to run
queries.
}
\description{
In order to establish we require the path in which the databases reside and
the name of the database to query.  Optionally you can specify a Java
classpath minimally including the ModelInterface.jar and BaseX.jar.  If no
classpath is given the version of ModelInterface and BaseX included in this
package will be used.
}
