% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querymi.R
\name{localDBConn}
\alias{localDBConn}
\title{Create a connection that can be used to run queries on a local GCAM database}
\usage{
localDBConn(dbPath, dbFile, miclasspath = NULL, migabble = TRUE,
  validatedb = TRUE)
}
\arguments{
\item{dbPath}{The path in which the BaseX DBs are located.}

\item{dbFile}{GCAM database to extract scenario from.}

\item{miclasspath}{Java class path for the GCAM Model Interface.}

\item{migabble}{If \code{TRUE}, discard model interface console output.  If
\code{FALSE}, display console output.}

\item{validatedb}{If \code{TRUE}, check that a simple db query works on the
connection; otherwise, don't run the check.}
}
\value{
A connection to a local BaseX databasse which can be used to run
queries.
}
\description{
Given a directory in which a GCAM database is located and the name of the
database, return a connection that can be used to run queries on the
database.
}
\details{
By default, the a version of the GCAM ModelInterface and BaseX libraries
supplied with the package will be used to run the query.  You can replace
these by specifying a Java classpath minimally including the replacement
ModelInterface.jar and BaseX.jar files.
}
