% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importgcam.R
\name{parse_mi_tables}
\alias{parse_mi_tables}
\title{Parse the GCAM ModelInterface output}
\usage{
parse_mi_tables(fn)
}
\arguments{
\item{fn}{Name of the file containing the output from the GCAM Model Interface.}
}
\description{
Parse the raw output of a GCAM batch query into a set of tables.
}

