% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querymi.R
\name{runQuery.localDBConn}
\alias{runQuery.localDBConn}
\title{Run query specialization for local databases}
\usage{
\method{runQuery}{localDBConn}(dbConn, query, scenarios, regions)
}
\arguments{
\item{dbConn}{The connection to a database which will handle running the query.}

\item{query}{A Model Interface query to run.  This is typically provided as the
XML typically found in the Main_queries.xml but could also be given for instance
as a query string that will result in XML such as:
\code{doc('../output/queries/Main_queries.xml')//*[@title='Cogeneration by sector']}}

\item{scenarios}{An array of scenarios to query. An empty list will imply the
last scenario in the database.  Each value can either be the scenario name or the
scenario name and date seperated by a space.  Note the date is specified exactly as
in appears in the database.}

\item{regions}{An array of regions to query. An empty list will imply "all-regions".}
}
\value{
A data.frame with the results.
}
\description{
Run query specialization for local databases
}
