% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querymi.R
\name{runQuery}
\alias{runQuery}
\title{Run queries via a database connection and recieve results as data.frame.}
\usage{
runQuery(dbConn, query, scenarios, regions)
}
\arguments{
\item{dbConn}{The connection to a database which will handle running the query.}

\item{query}{A Model Interface query to run.  This is typically provided as the
XML typically found in the Main_queries.xml but could also be given for instance
as a query string that will result in XML such as:
\code{collection('../output/queries/Main_queries.xml')//*[@title='Cogeneration by sector']}}

\item{scenarios}{An array of scenarios to query. TODO: more about syntax}

\item{regions}{An array of regions to query. TODO: allow null to mean all?}
}
\value{
A data.frame with the results.  TODO: what about errors?
}
\description{
To run a query users need to supply a database connection, the query to run,
the scenarios to query, and regions to query.
}
