% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageProjectData.R
\name{saveProject}
\alias{saveProject}
\title{Save project data to a backing file.}
\usage{
saveProject(prjdata, file = NULL)
}
\arguments{
\item{prjdata}{Project data object.}

\item{file}{Filename to save to.  If \code{NULL}, use the file the project
was loaded from.}
}
\description{
Write a project data structure to its data file.  This should rarely be
necessary, since most of the functions that modify a project file have an
option to write the data back to the file.  Mostly it is intended for
internal use, but it's exported in case it's useful to someone for some
reason.
}

